[build] Fix unused variable warnings and casting warnings

+++ src/CameraLocatorEntity.cpp
@@ -42,7 +42,7 @@ CameraLocatorEntity::CameraLocatorEntity(Qt3DCore::QNode* parent)
         };
 
 
-    QByteArray positionData((const char*)points.data(), points.size() * sizeof(float));
+    QByteArray positionData((const char*)points.data(), points.size() * static_cast<int>(sizeof(float)));
     auto vertexDataBuffer = new QBuffer;
     vertexDataBuffer->setData(positionData);
     auto positionAttribute = new QAttribute;
@@ -52,7 +52,7 @@ CameraLocatorEntity::CameraLocatorEntity(Qt3DCore::QNode* parent)
     positionAttribute->setVertexSize(3);
     positionAttribute->setByteOffset(0);
     positionAttribute->setByteStride(3 * sizeof(float));
-    positionAttribute->setCount(points.size() / 3);
+    positionAttribute->setCount(static_cast<uint>(points.size() / 3));
     positionAttribute->setName(QAttribute::defaultPositionAttributeName());
     customGeometry->addAttribute(positionAttribute);
 
@@ -77,7 +77,7 @@ CameraLocatorEntity::CameraLocatorEntity(Qt3DCore::QNode* parent)
         1.f, 1.f, 1.f, 1.f, 1.f, 1.f,
         };
 
-    QByteArray colorData((const char*)colors.data(), colors.size() * sizeof(float));
+    QByteArray colorData((const char*)colors.data(), colors.size() * static_cast<int>(sizeof(float)));
     auto colorDataBuffer = new QBuffer;
     colorDataBuffer->setData(colorData);
     auto colorAttribute = new QAttribute;
@@ -87,7 +87,7 @@ CameraLocatorEntity::CameraLocatorEntity(Qt3DCore::QNode* parent)
     colorAttribute->setVertexSize(3);
     colorAttribute->setByteOffset(0);
     colorAttribute->setByteStride(3 * sizeof(float));
-    colorAttribute->setCount(colors.size() / 3);
+    colorAttribute->setCount(static_cast<uint>(colors.size() / 3));
     colorAttribute->setName(QAttribute::defaultColorAttributeName());
     customGeometry->addAttribute(colorAttribute);
